rails4 magic sauce

+++ db/migrate/20121123054127_make_post_number_distinct.rb
@@ -20,8 +20,11 @@ from
 ) as X
 where calc <> p.post_number and X.id = p.id')
 
-    remove_index :posts, :forum_thread_id_and_post_number
-    add_index :posts, [:topic_id, :post_number], unique: true
+    # automagically does this
+    unless rails4?
+      remove_index :posts, :forum_thread_id_and_post_number
+      add_index :posts, [:topic_id, :post_number], unique: true
+    end
   end
 
   def down
